all: rename the project and the engine
The engine is renamed to "pkcs11" because that is what wpa_supplicant expects.

The engine file is renamed from libp11-kit.so, because libtool will confuse
that with the actual -lp11-kit.

The project is renamed to reflect that it will include the rest of the
functionality to make PKCS#11 remoting work.